main/requirements.txt
fastapi
uvicorn
sqlalchemy
Pillow
python-multipart
transformers
requests
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio